implement parsing for `UciResponse`
resolves #1

Note: This commit extends the impact of #4. The relevant codepath
mentioned there is now present regardless of the `types` feature.
As a consequence, compilation currently always fails with the
`validate-promotion-moves` feature

Not a breaking change, the shape of `UciResponse` remains unchanged.
This required some gymnastics in particular when parsing uci moves:
`UciCommand` contained `UciMove`s instead of `String`s with the `types`
feature enabled and the parsing infrastructure was specialized for this.
Since `UciResponse` stores uci moves as `String` even with the `types`
feature, the alternative parsing functions `recognize_uci_move`,
`recognize_moves` and `recognize_moves_after` were created to allow
always parsing into `String`.

Tests are a mix of made-up and grabbed from the UCI spec